kfence: Cleanup KFENCE-internal interface
In preparation of factoring out reporting code into report.c, rename
core.h -> kfence.h, and apply trivial cleanups.
